+++ tiger-common/src/main/java/de/gematik/test/tiger/common/config/BasicTigerConfigurationSource.java
@@ -10,6 +10,9 @@ import java.util.Map;
 import java.util.stream.Collectors;
 import lombok.Builder;
 
+/**
+ * Stores a map of key/value-pairs.
+ */
 public class BasicTigerConfigurationSource extends AbstractTigerConfigurationSource {
 
     private final Map<TigerConfigurationKey, String> values;